@@ -15,29 +15,29 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 # In iOS, build-name is used as CFBundleShortVersionString while build-number used as CFBundleVersion.
 # Read more about iOS versioning at
 # https://developer.apple.com/library/archive/documentation/General/Reference/InfoPlistKeyReference/Articles/CoreFoundationKeys.html
-version: 1.0.4+51
+version: 1.0.4+55
 
 environment:
-  sdk: '>=2.11.0 <3.0.0'
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
-  auto_route: ^0.6.9
-  firebase_crashlytics: ^1.0.0
-  get_it: ^5.0.6
-  google_mobile_ads: ^0.11.0+4
-  hue_dart: ^0.2.1
-  in_app_purchase: ^0.3.5+1
-  in_app_review: ^2.0.1
-  injectable:
-  introduction_screen: ^2.0.0
-  json_annotation: ^3.1.1
-  logger: ^0.9.4
-  material_design_icons_flutter: ^4.0.5955
-  provider: ^4.3.3
-  shared_preferences: ^2.0.4
-  stacked: ^1.9.5
-  stacked_services: ^0.7.3
-  url_launcher: ^6.0.2
+  collection: ^1.15.0
+  firebase_crashlytics: ^2.0.0
+  get_it: ^6.1.0
+  google_mobile_ads: ^0.13.0
+  hue_dart: ^0.3.0
+  in_app_purchase: ^0.5.0
+  in_app_review: ^2.0.0
+  injectable: ^1.3.0
+  introduction_screen: ^2.1.0
+  json_annotation: ^4.0.0
+  logger: ^1.0.0
+  material_design_icons_flutter: ^5.0.5955-rc.1
+  provider: ^5.0.0
+  shared_preferences: ^2.0.0
+  stacked: ^2.1.0
+  stacked_services: ^0.8.0
+  url_launcher: ^6.0.0
   
   device_functions:
     path: ../device_functions/
@@ -52,17 +52,21 @@ dependencies:
     sdk: flutter
 
 dependency_overrides:
+  clock: ^1.1.0
+  http: ^0.13.3
   intl: ^0.17.0
   path: ^1.8.0
-  clock: ^1.1.0
 
 dev_dependencies:
-  auto_route_generator: ^0.6.10
   build_runner:
-  injectable_generator:
-  json_serializable: ^3.5.1
-  pedantic:
-  pigeon: ^0.1.23
+  injectable_generator: ^1.3.0
+  json_serializable: ^4.1.0
+  pedantic: ^1.11.0
+  pigeon: ^0.2.0
+  stacked_generator: ^0.4.0
+
+  mockito: ^5.0.0
+  test: ^1.17.0
 
 flutter_intl:
   enabled: true
