@@ -7,7 +7,7 @@ class MyTheme {
   final ThemeData theme;
 
   const MyTheme(
-      {@required this.id, @required this.title, @required this.theme});
+      {required this.id, required this.title, required this.theme});
 }
 
 const _lightFillColor = Colors.black;
@@ -103,7 +103,7 @@ ThemeData _themeData(ColorScheme colorScheme, Color focusColor) {
         _lightFillColor.withOpacity(0.80),
         _darkFillColor,
       ),
-      contentTextStyle: _textTheme.subtitle1.apply(color: _darkFillColor),
+      contentTextStyle: _textTheme.subtitle1!.apply(color: _darkFillColor),
     ),
     toggleButtonsTheme: ToggleButtonsThemeData(
       constraints: BoxConstraints(minWidth: 80),
@@ -119,13 +119,13 @@ ThemeData _themeData(ColorScheme colorScheme, Color focusColor) {
 }
 
 ColorScheme _lightOrangeColorScheme = ColorScheme(
-  primary: Colors.blueGrey[600],
-  primaryVariant: Colors.deepOrangeAccent[100],
-  secondary: Colors.deepOrangeAccent[100],
-  secondaryVariant: Colors.deepOrange[300],
+  primary: Colors.blueGrey[600]!,
+  primaryVariant: Colors.deepOrangeAccent[100]!,
+  secondary: Colors.deepOrangeAccent[100]!,
+  secondaryVariant: Colors.deepOrange[300]!,
   background: Color(0xFFFCFCFC),
   surface: Color(0xFFF4F4F4),
-  onBackground: Colors.grey[900],
+  onBackground: Colors.grey[900]!,
   error: _lightFillColor,
   onError: _lightFillColor,
   onPrimary: _lightFillColor,
@@ -135,13 +135,13 @@ ColorScheme _lightOrangeColorScheme = ColorScheme(
 );
 
 ColorScheme _lightGreenColorScheme = ColorScheme(
-  primary: Colors.blueGrey[600],
+  primary: Colors.blueGrey[600]!,
   primaryVariant: Color(0xFF68D391),
   secondary: Color(0xFF68D391),
   secondaryVariant: Color(0xFF48BB78),
   background: Color(0xFFFCFCFC),
   surface: Color(0xFFF4F4F4),
-  onBackground: Colors.grey[900],
+  onBackground: Colors.grey[900]!,
   error: _lightFillColor,
   onError: _lightFillColor,
   onPrimary: _lightFillColor,
@@ -151,7 +151,7 @@ ColorScheme _lightGreenColorScheme = ColorScheme(
 );
 
 ColorScheme _darkOrangeColorScheme = ColorScheme(
-  primary: Colors.deepOrange[300],
+  primary: Colors.deepOrange[300]!,
   primaryVariant: Colors.tealAccent,
   secondary: Colors.deepPurpleAccent,
   secondaryVariant: Colors.deepPurple,
@@ -167,10 +167,10 @@ ColorScheme _darkOrangeColorScheme = ColorScheme(
 );
 
 ColorScheme _darkYellowColorScheme = ColorScheme(
-  primary: Colors.yellow[800],
-  primaryVariant: Colors.tealAccent[400],
-  secondary: Colors.teal[700],
-  secondaryVariant: Colors.teal[800],
+  primary: Colors.yellow[800]!,
+  primaryVariant: Colors.tealAccent[400]!,
+  secondary: Colors.teal[700]!,
+  secondaryVariant: Colors.teal[800]!,
   background: Color(0xFF1b1b1b),
   surface: Color(0xFF373737),
   onBackground: Color(0x0DFFFFFF),
@@ -183,10 +183,10 @@ ColorScheme _darkYellowColorScheme = ColorScheme(
 );
 
 ColorScheme _blackColorScheme = ColorScheme(
-  primary: Colors.indigo[200],
-  primaryVariant: Colors.indigoAccent[100],
-  secondary: Colors.indigo[500],
-  secondaryVariant: Colors.indigo[400],
+  primary: Colors.indigo[200]!,
+  primaryVariant: Colors.indigoAccent[100]!,
+  secondary: Colors.indigo[500]!,
+  secondaryVariant: Colors.indigo[400]!,
   background: Colors.black,
   surface: Color(0xFF1A1A1A),
   onBackground: Color(0x0DFFFFFF),
