@@ -1,149 +1,149 @@
-// Autogenerated from Pigeon (v0.1.23), do not edit directly.
+// Autogenerated from Pigeon (v0.2.1), do not edit directly.
 // See also: https://pub.dev/packages/pigeon
-// ignore_for_file: public_member_api_docs, non_constant_identifier_names, avoid_as, unused_import, unnecessary_parenthesis, prefer_null_aware_operators
-// @dart = 2.8
+// ignore_for_file: public_member_api_docs, non_constant_identifier_names, avoid_as, unused_import, unnecessary_parenthesis, prefer_null_aware_operators, omit_local_variable_types
+// @dart = 2.12
 import 'dart:async';
 import 'dart:typed_data' show Uint8List, Int32List, Int64List, Float64List;
 
 import 'package:flutter/services.dart';
 
-class ExtendTimeResponse {
-  String timerId;
-  int additionalTime;
+class ExtendTimeRequest {
+  String? timerId;
+  int? additionalTime;
 
   Object encode() {
-    final Map<Object, Object> pigeonMap = <Object, Object>{};
+    final Map<Object?, Object?> pigeonMap = <Object?, Object?>{};
     pigeonMap['timerId'] = timerId;
     pigeonMap['additionalTime'] = additionalTime;
     return pigeonMap;
   }
 
-  static ExtendTimeResponse decode(Object message) {
-    final Map<Object, Object> pigeonMap = message as Map<Object, Object>;
-    return ExtendTimeResponse()
-      ..timerId = pigeonMap['timerId'] as String
-      ..additionalTime = pigeonMap['additionalTime'] as int;
+  static ExtendTimeRequest decode(Object message) {
+    final Map<Object?, Object?> pigeonMap = message as Map<Object?, Object?>;
+    return ExtendTimeRequest()
+      ..timerId = pigeonMap['timerId'] as String?
+      ..additionalTime = pigeonMap['additionalTime'] as int?;
   }
 }
 
 class CountDownRequest {
-  String timerId;
-  int newTime;
+  String? timerId;
+  int? newTime;
 
   Object encode() {
-    final Map<Object, Object> pigeonMap = <Object, Object>{};
+    final Map<Object?, Object?> pigeonMap = <Object?, Object?>{};
     pigeonMap['timerId'] = timerId;
     pigeonMap['newTime'] = newTime;
     return pigeonMap;
   }
 
   static CountDownRequest decode(Object message) {
-    final Map<Object, Object> pigeonMap = message as Map<Object, Object>;
+    final Map<Object?, Object?> pigeonMap = message as Map<Object?, Object?>;
     return CountDownRequest()
-      ..timerId = pigeonMap['timerId'] as String
-      ..newTime = pigeonMap['newTime'] as int;
+      ..timerId = pigeonMap['timerId'] as String?
+      ..newTime = pigeonMap['newTime'] as int?;
   }
 }
 
 class TimerRequest {
-  String timerId;
+  String? timerId;
 
   Object encode() {
-    final Map<Object, Object> pigeonMap = <Object, Object>{};
+    final Map<Object?, Object?> pigeonMap = <Object?, Object?>{};
     pigeonMap['timerId'] = timerId;
     return pigeonMap;
   }
 
   static TimerRequest decode(Object message) {
-    final Map<Object, Object> pigeonMap = message as Map<Object, Object>;
+    final Map<Object?, Object?> pigeonMap = message as Map<Object?, Object?>;
     return TimerRequest()
-      ..timerId = pigeonMap['timerId'] as String;
+      ..timerId = pigeonMap['timerId'] as String?;
   }
 }
 
 class OpenRequest {
-  String timerId;
+  String? timerId;
 
   Object encode() {
-    final Map<Object, Object> pigeonMap = <Object, Object>{};
+    final Map<Object?, Object?> pigeonMap = <Object?, Object?>{};
     pigeonMap['timerId'] = timerId;
     return pigeonMap;
   }
 
   static OpenRequest decode(Object message) {
-    final Map<Object, Object> pigeonMap = message as Map<Object, Object>;
+    final Map<Object?, Object?> pigeonMap = message as Map<Object?, Object?>;
     return OpenRequest()
-      ..timerId = pigeonMap['timerId'] as String;
+      ..timerId = pigeonMap['timerId'] as String?;
   }
 }
 
 class WidgetUpdateResponse {
-  String title;
+  String? title;
 
   Object encode() {
-    final Map<Object, Object> pigeonMap = <Object, Object>{};
+    final Map<Object?, Object?> pigeonMap = <Object?, Object?>{};
     pigeonMap['title'] = title;
     return pigeonMap;
   }
 
   static WidgetUpdateResponse decode(Object message) {
-    final Map<Object, Object> pigeonMap = message as Map<Object, Object>;
+    final Map<Object?, Object?> pigeonMap = message as Map<Object?, Object?>;
     return WidgetUpdateResponse()
-      ..title = pigeonMap['title'] as String;
+      ..title = pigeonMap['title'] as String?;
   }
 }
 
 class InitializationRequest {
-  int callbackHandle;
+  int? callbackHandle;
 
   Object encode() {
-    final Map<Object, Object> pigeonMap = <Object, Object>{};
+    final Map<Object?, Object?> pigeonMap = <Object?, Object?>{};
     pigeonMap['callbackHandle'] = callbackHandle;
     return pigeonMap;
   }
 
   static InitializationRequest decode(Object message) {
-    final Map<Object, Object> pigeonMap = message as Map<Object, Object>;
+    final Map<Object?, Object?> pigeonMap = message as Map<Object?, Object?>;
     return InitializationRequest()
-      ..callbackHandle = pigeonMap['callbackHandle'] as int;
+      ..callbackHandle = pigeonMap['callbackHandle'] as int?;
   }
 }
 
 class NotificationResponse {
-  String timerId;
-  bool success;
+  String? timerId;
+  bool? success;
 
   Object encode() {
-    final Map<Object, Object> pigeonMap = <Object, Object>{};
+    final Map<Object?, Object?> pigeonMap = <Object?, Object?>{};
     pigeonMap['timerId'] = timerId;
     pigeonMap['success'] = success;
     return pigeonMap;
   }
 
   static NotificationResponse decode(Object message) {
-    final Map<Object, Object> pigeonMap = message as Map<Object, Object>;
+    final Map<Object?, Object?> pigeonMap = message as Map<Object?, Object?>;
     return NotificationResponse()
-      ..timerId = pigeonMap['timerId'] as String
-      ..success = pigeonMap['success'] as bool;
+      ..timerId = pigeonMap['timerId'] as String?
+      ..success = pigeonMap['success'] as bool?;
   }
 }
 
 class RunningNotificationRequest {
-  String description;
-  String timerId;
-  String title;
-  int accentColor;
-  String restartAction;
-  String continueAction;
-  String pauseAction;
-  String cancelAction;
-  List<Object> extendActions;
-  int duration;
-  int remainingTime;
-  bool shakeToExtend;
+  String? description;
+  String? timerId;
+  String? title;
+  int? accentColor;
+  String? restartAction;
+  String? continueAction;
+  String? pauseAction;
+  String? cancelAction;
+  List<Object?>? extendActions;
+  int? duration;
+  int? remainingTime;
+  bool? shakeToExtend;
 
   Object encode() {
-    final Map<Object, Object> pigeonMap = <Object, Object>{};
+    final Map<Object?, Object?> pigeonMap = <Object?, Object?>{};
     pigeonMap['description'] = description;
     pigeonMap['timerId'] = timerId;
     pigeonMap['title'] = title;
@@ -160,38 +160,38 @@ class RunningNotificationRequest {
   }
 
   static RunningNotificationRequest decode(Object message) {
-    final Map<Object, Object> pigeonMap = message as Map<Object, Object>;
+    final Map<Object?, Object?> pigeonMap = message as Map<Object?, Object?>;
     return RunningNotificationRequest()
-      ..description = pigeonMap['description'] as String
-      ..timerId = pigeonMap['timerId'] as String
-      ..title = pigeonMap['title'] as String
-      ..accentColor = pigeonMap['accentColor'] as int
-      ..restartAction = pigeonMap['restartAction'] as String
-      ..continueAction = pigeonMap['continueAction'] as String
-      ..pauseAction = pigeonMap['pauseAction'] as String
-      ..cancelAction = pigeonMap['cancelAction'] as String
-      ..extendActions = pigeonMap['extendActions'] as List<Object>
-      ..duration = pigeonMap['duration'] as int
-      ..remainingTime = pigeonMap['remainingTime'] as int
-      ..shakeToExtend = pigeonMap['shakeToExtend'] as bool;
+      ..description = pigeonMap['description'] as String?
+      ..timerId = pigeonMap['timerId'] as String?
+      ..title = pigeonMap['title'] as String?
+      ..accentColor = pigeonMap['accentColor'] as int?
+      ..restartAction = pigeonMap['restartAction'] as String?
+      ..continueAction = pigeonMap['continueAction'] as String?
+      ..pauseAction = pigeonMap['pauseAction'] as String?
+      ..cancelAction = pigeonMap['cancelAction'] as String?
+      ..extendActions = pigeonMap['extendActions'] as List<Object?>?
+      ..duration = pigeonMap['duration'] as int?
+      ..remainingTime = pigeonMap['remainingTime'] as int?
+      ..shakeToExtend = pigeonMap['shakeToExtend'] as bool?;
   }
 }
 
 class TimeNotificationRequest {
-  String description;
-  String timerId;
-  String title;
-  int accentColor;
-  String restartAction;
-  String continueAction;
-  String pauseAction;
-  String cancelAction;
-  List<Object> extendActions;
-  int duration;
-  int remainingTime;
+  String? description;
+  String? timerId;
+  String? title;
+  int? accentColor;
+  String? restartAction;
+  String? continueAction;
+  String? pauseAction;
+  String? cancelAction;
+  List<Object?>? extendActions;
+  int? duration;
+  int? remainingTime;
 
   Object encode() {
-    final Map<Object, Object> pigeonMap = <Object, Object>{};
+    final Map<Object?, Object?> pigeonMap = <Object?, Object?>{};
     pigeonMap['description'] = description;
     pigeonMap['timerId'] = timerId;
     pigeonMap['title'] = title;
@@ -207,35 +207,35 @@ class TimeNotificationRequest {
   }
 
   static TimeNotificationRequest decode(Object message) {
-    final Map<Object, Object> pigeonMap = message as Map<Object, Object>;
+    final Map<Object?, Object?> pigeonMap = message as Map<Object?, Object?>;
     return TimeNotificationRequest()
-      ..description = pigeonMap['description'] as String
-      ..timerId = pigeonMap['timerId'] as String
-      ..title = pigeonMap['title'] as String
-      ..accentColor = pigeonMap['accentColor'] as int
-      ..restartAction = pigeonMap['restartAction'] as String
-      ..continueAction = pigeonMap['continueAction'] as String
-      ..pauseAction = pigeonMap['pauseAction'] as String
-      ..cancelAction = pigeonMap['cancelAction'] as String
-      ..extendActions = pigeonMap['extendActions'] as List<Object>
-      ..duration = pigeonMap['duration'] as int
-      ..remainingTime = pigeonMap['remainingTime'] as int;
+      ..description = pigeonMap['description'] as String?
+      ..timerId = pigeonMap['timerId'] as String?
+      ..title = pigeonMap['title'] as String?
+      ..accentColor = pigeonMap['accentColor'] as int?
+      ..restartAction = pigeonMap['restartAction'] as String?
+      ..continueAction = pigeonMap['continueAction'] as String?
+      ..pauseAction = pigeonMap['pauseAction'] as String?
+      ..cancelAction = pigeonMap['cancelAction'] as String?
+      ..extendActions = pigeonMap['extendActions'] as List<Object?>?
+      ..duration = pigeonMap['duration'] as int?
+      ..remainingTime = pigeonMap['remainingTime'] as int?;
   }
 }
 
 class NotificationRequest {
-  String timerId;
-  String title;
-  String description;
-  int accentColor;
-  String restartAction;
-  String continueAction;
-  String pauseAction;
-  String cancelAction;
-  List<Object> extendActions;
+  String? timerId;
+  String? title;
+  String? description;
+  int? accentColor;
+  String? restartAction;
+  String? continueAction;
+  String? pauseAction;
+  String? cancelAction;
+  List<Object?>? extendActions;
 
   Object encode() {
-    final Map<Object, Object> pigeonMap = <Object, Object>{};
+    final Map<Object?, Object?> pigeonMap = <Object?, Object?>{};
     pigeonMap['timerId'] = timerId;
     pigeonMap['title'] = title;
     pigeonMap['description'] = description;
@@ -249,94 +249,110 @@ class NotificationRequest {
   }
 
   static NotificationRequest decode(Object message) {
-    final Map<Object, Object> pigeonMap = message as Map<Object, Object>;
+    final Map<Object?, Object?> pigeonMap = message as Map<Object?, Object?>;
     return NotificationRequest()
-      ..timerId = pigeonMap['timerId'] as String
-      ..title = pigeonMap['title'] as String
-      ..description = pigeonMap['description'] as String
-      ..accentColor = pigeonMap['accentColor'] as int
-      ..restartAction = pigeonMap['restartAction'] as String
-      ..continueAction = pigeonMap['continueAction'] as String
-      ..pauseAction = pigeonMap['pauseAction'] as String
-      ..cancelAction = pigeonMap['cancelAction'] as String
-      ..extendActions = pigeonMap['extendActions'] as List<Object>;
+      ..timerId = pigeonMap['timerId'] as String?
+      ..title = pigeonMap['title'] as String?
+      ..description = pigeonMap['description'] as String?
+      ..accentColor = pigeonMap['accentColor'] as int?
+      ..restartAction = pigeonMap['restartAction'] as String?
+      ..continueAction = pigeonMap['continueAction'] as String?
+      ..pauseAction = pigeonMap['pauseAction'] as String?
+      ..cancelAction = pigeonMap['cancelAction'] as String?
+      ..extendActions = pigeonMap['extendActions'] as List<Object?>?;
   }
 }
 
 class CancelResponse {
-  String timerId;
-  bool success;
+  String? timerId;
+  bool? success;
 
   Object encode() {
-    final Map<Object, Object> pigeonMap = <Object, Object>{};
+    final Map<Object?, Object?> pigeonMap = <Object?, Object?>{};
     pigeonMap['timerId'] = timerId;
     pigeonMap['success'] = success;
     return pigeonMap;
   }
 
   static CancelResponse decode(Object message) {
-    final Map<Object, Object> pigeonMap = message as Map<Object, Object>;
+    final Map<Object?, Object?> pigeonMap = message as Map<Object?, Object?>;
     return CancelResponse()
-      ..timerId = pigeonMap['timerId'] as String
-      ..success = pigeonMap['success'] as bool;
+      ..timerId = pigeonMap['timerId'] as String?
+      ..success = pigeonMap['success'] as bool?;
   }
 }
 
 class CancelRequest {
-  String timerId;
+  String? timerId;
 
   Object encode() {
-    final Map<Object, Object> pigeonMap = <Object, Object>{};
+    final Map<Object?, Object?> pigeonMap = <Object?, Object?>{};
     pigeonMap['timerId'] = timerId;
     return pigeonMap;
   }
 
   static CancelRequest decode(Object message) {
-    final Map<Object, Object> pigeonMap = message as Map<Object, Object>;
+    final Map<Object?, Object?> pigeonMap = message as Map<Object?, Object?>;
     return CancelRequest()
-      ..timerId = pigeonMap['timerId'] as String;
+      ..timerId = pigeonMap['timerId'] as String?;
+  }
+}
+
+class ToggleRequest {
+  bool? enable;
+
+  Object encode() {
+    final Map<Object?, Object?> pigeonMap = <Object?, Object?>{};
+    pigeonMap['enable'] = enable;
+    return pigeonMap;
+  }
+
+  static ToggleRequest decode(Object message) {
+    final Map<Object?, Object?> pigeonMap = message as Map<Object?, Object?>;
+    return ToggleRequest()
+      ..enable = pigeonMap['enable'] as bool?;
   }
 }
 
 class InstalledAppsResponse {
-  List<Object> apps;
+  List<Object?>? apps;
 
   Object encode() {
-    final Map<Object, Object> pigeonMap = <Object, Object>{};
+    final Map<Object?, Object?> pigeonMap = <Object?, Object?>{};
     pigeonMap['apps'] = apps;
     return pigeonMap;
   }
 
   static InstalledAppsResponse decode(Object message) {
-    final Map<Object, Object> pigeonMap = message as Map<Object, Object>;
+    final Map<Object?, Object?> pigeonMap = message as Map<Object?, Object?>;
     return InstalledAppsResponse()
-      ..apps = pigeonMap['apps'] as List<Object>;
+      ..apps = pigeonMap['apps'] as List<Object?>?;
   }
 }
 
 class LaunchAppRequest {
-  String packageName;
+  String? packageName;
 
   Object encode() {
-    final Map<Object, Object> pigeonMap = <Object, Object>{};
+    final Map<Object?, Object?> pigeonMap = <Object?, Object?>{};
     pigeonMap['packageName'] = packageName;
     return pigeonMap;
   }
 
   static LaunchAppRequest decode(Object message) {
-    final Map<Object, Object> pigeonMap = message as Map<Object, Object>;
+    final Map<Object?, Object?> pigeonMap = message as Map<Object?, Object?>;
     return LaunchAppRequest()
-      ..packageName = pigeonMap['packageName'] as String;
+      ..packageName = pigeonMap['packageName'] as String?;
   }
 }
 
 class Package {
-  String title;
-  String icon;
-  String packageName;
+  String? title;
+  String? icon;
+  String? packageName;
 
   Object encode() {
-    final Map<Object, Object> pigeonMap = <Object, Object>{};
+    final Map<Object?, Object?> pigeonMap = <Object?, Object?>{};
     pigeonMap['title'] = title;
     pigeonMap['icon'] = icon;
     pigeonMap['packageName'] = packageName;
@@ -344,16 +360,16 @@ class Package {
   }
 
   static Package decode(Object message) {
-    final Map<Object, Object> pigeonMap = message as Map<Object, Object>;
+    final Map<Object?, Object?> pigeonMap = message as Map<Object?, Object?>;
     return Package()
-      ..title = pigeonMap['title'] as String
-      ..icon = pigeonMap['icon'] as String
-      ..packageName = pigeonMap['packageName'] as String;
+      ..title = pigeonMap['title'] as String?
+      ..icon = pigeonMap['icon'] as String?
+      ..packageName = pigeonMap['packageName'] as String?;
   }
 }
 
 abstract class FlutterTimerApi {
-  void onExtendTime(ExtendTimeResponse arg);
+  void onExtendTime(ExtendTimeRequest arg);
   void onCountDown(CountDownRequest arg);
   void onContinueRequest(TimerRequest arg);
   void onPauseRequest(TimerRequest arg);
@@ -363,126 +379,126 @@ abstract class FlutterTimerApi {
   void onAlarm(TimerRequest arg);
   WidgetUpdateResponse onWidgetUpdate();
   void onWidgetStartTimer();
-  static void setup(FlutterTimerApi api) {
+  static void setup(FlutterTimerApi? api) {
     {
-      const BasicMessageChannel<Object> channel =
-          BasicMessageChannel<Object>('dev.flutter.pigeon.FlutterTimerApi.onExtendTime', StandardMessageCodec());
+      const BasicMessageChannel<Object?> channel = BasicMessageChannel<Object?>(
+          'dev.flutter.pigeon.FlutterTimerApi.onExtendTime', StandardMessageCodec());
       if (api == null) {
         channel.setMessageHandler(null);
       } else {
-        channel.setMessageHandler((Object message) async {
-          assert(message != null, 'Argument for dev.flutter.pigeon.FlutterTimerApi.onExtendTime was null. Expected ExtendTimeResponse.');
-          final ExtendTimeResponse input = ExtendTimeResponse.decode(message);
+        channel.setMessageHandler((Object? message) async {
+          assert(message != null, 'Argument for dev.flutter.pigeon.FlutterTimerApi.onExtendTime was null. Expected ExtendTimeRequest.');
+          final ExtendTimeRequest input = ExtendTimeRequest.decode(message!);
           api.onExtendTime(input);
           return;
         });
       }
     }
     {
-      const BasicMessageChannel<Object> channel =
-          BasicMessageChannel<Object>('dev.flutter.pigeon.FlutterTimerApi.onCountDown', StandardMessageCodec());
+      const BasicMessageChannel<Object?> channel = BasicMessageChannel<Object?>(
+          'dev.flutter.pigeon.FlutterTimerApi.onCountDown', StandardMessageCodec());
       if (api == null) {
         channel.setMessageHandler(null);
       } else {
-        channel.setMessageHandler((Object message) async {
+        channel.setMessageHandler((Object? message) async {
           assert(message != null, 'Argument for dev.flutter.pigeon.FlutterTimerApi.onCountDown was null. Expected CountDownRequest.');
-          final CountDownRequest input = CountDownRequest.decode(message);
+          final CountDownRequest input = CountDownRequest.decode(message!);
           api.onCountDown(input);
           return;
         });
       }
     }
     {
-      const BasicMessageChannel<Object> channel =
-          BasicMessageChannel<Object>('dev.flutter.pigeon.FlutterTimerApi.onContinueRequest', StandardMessageCodec());
+      const BasicMessageChannel<Object?> channel = BasicMessageChannel<Object?>(
+          'dev.flutter.pigeon.FlutterTimerApi.onContinueRequest', StandardMessageCodec());
       if (api == null) {
         channel.setMessageHandler(null);
       } else {
-        channel.setMessageHandler((Object message) async {
+        channel.setMessageHandler((Object? message) async {
           assert(message != null, 'Argument for dev.flutter.pigeon.FlutterTimerApi.onContinueRequest was null. Expected TimerRequest.');
-          final TimerRequest input = TimerRequest.decode(message);
+          final TimerRequest input = TimerRequest.decode(message!);
           api.onContinueRequest(input);
           return;
         });
       }
     }
     {
-      const BasicMessageChannel<Object> channel =
-          BasicMessageChannel<Object>('dev.flutter.pigeon.FlutterTimerApi.onPauseRequest', StandardMessageCodec());
+      const BasicMessageChannel<Object?> channel = BasicMessageChannel<Object?>(
+          'dev.flutter.pigeon.FlutterTimerApi.onPauseRequest', StandardMessageCodec());
       if (api == null) {
         channel.setMessageHandler(null);
       } else {
-        channel.setMessageHandler((Object message) async {
+        channel.setMessageHandler((Object? message) async {
           assert(message != null, 'Argument for dev.flutter.pigeon.FlutterTimerApi.onPauseRequest was null. Expected TimerRequest.');
-          final TimerRequest input = TimerRequest.decode(message);
+          final TimerRequest input = TimerRequest.decode(message!);
           api.onPauseRequest(input);
           return;
         });
       }
     }
     {
-      const BasicMessageChannel<Object> channel =
-          BasicMessageChannel<Object>('dev.flutter.pigeon.FlutterTimerApi.onCancelRequest', StandardMessageCodec());
+      const BasicMessageChannel<Object?> channel = BasicMessageChannel<Object?>(
+          'dev.flutter.pigeon.FlutterTimerApi.onCancelRequest', StandardMessageCodec());
       if (api == null) {
         channel.setMessageHandler(null);
       } else {
-        channel.setMessageHandler((Object message) async {
+        channel.setMessageHandler((Object? message) async {
           assert(message != null, 'Argument for dev.flutter.pigeon.FlutterTimerApi.onCancelRequest was null. Expected TimerRequest.');
-          final TimerRequest input = TimerRequest.decode(message);
+          final TimerRequest input = TimerRequest.decode(message!);
           api.onCancelRequest(input);
           return;
         });
       }
     }
     {
-      const BasicMessageChannel<Object> channel =
-          BasicMessageChannel<Object>('dev.flutter.pigeon.FlutterTimerApi.onRestartRequest', StandardMessageCodec());
+      const BasicMessageChannel<Object?> channel = BasicMessageChannel<Object?>(
+          'dev.flutter.pigeon.FlutterTimerApi.onRestartRequest', StandardMessageCodec());
       if (api == null) {
         channel.setMessageHandler(null);
       } else {
-        channel.setMessageHandler((Object message) async {
+        channel.setMessageHandler((Object? message) async {
           assert(message != null, 'Argument for dev.flutter.pigeon.FlutterTimerApi.onRestartRequest was null. Expected TimerRequest.');
-          final TimerRequest input = TimerRequest.decode(message);
+          final TimerRequest input = TimerRequest.decode(message!);
           api.onRestartRequest(input);
           return;
         });
       }
     }
     {
-      const BasicMessageChannel<Object> channel =
-          BasicMessageChannel<Object>('dev.flutter.pigeon.FlutterTimerApi.onOpen', StandardMessageCodec());
+      const BasicMessageChannel<Object?> channel = BasicMessageChannel<Object?>(
+          'dev.flutter.pigeon.FlutterTimerApi.onOpen', StandardMessageCodec());
       if (api == null) {
         channel.setMessageHandler(null);
       } else {
-        channel.setMessageHandler((Object message) async {
+        channel.setMessageHandler((Object? message) async {
           assert(message != null, 'Argument for dev.flutter.pigeon.FlutterTimerApi.onOpen was null. Expected OpenRequest.');
-          final OpenRequest input = OpenRequest.decode(message);
+          final OpenRequest input = OpenRequest.decode(message!);
           api.onOpen(input);
           return;
         });
       }
     }
     {
-      const BasicMessageChannel<Object> channel =
-          BasicMessageChannel<Object>('dev.flutter.pigeon.FlutterTimerApi.onAlarm', StandardMessageCodec());
+      const BasicMessageChannel<Object?> channel = BasicMessageChannel<Object?>(
+          'dev.flutter.pigeon.FlutterTimerApi.onAlarm', StandardMessageCodec());
       if (api == null) {
         channel.setMessageHandler(null);
       } else {
-        channel.setMessageHandler((Object message) async {
+        channel.setMessageHandler((Object? message) async {
           assert(message != null, 'Argument for dev.flutter.pigeon.FlutterTimerApi.onAlarm was null. Expected TimerRequest.');
-          final TimerRequest input = TimerRequest.decode(message);
+          final TimerRequest input = TimerRequest.decode(message!);
           api.onAlarm(input);
           return;
         });
       }
     }
     {
-      const BasicMessageChannel<Object> channel =
-          BasicMessageChannel<Object>('dev.flutter.pigeon.FlutterTimerApi.onWidgetUpdate', StandardMessageCodec());
+      const BasicMessageChannel<Object?> channel = BasicMessageChannel<Object?>(
+          'dev.flutter.pigeon.FlutterTimerApi.onWidgetUpdate', StandardMessageCodec());
       if (api == null) {
         channel.setMessageHandler(null);
       } else {
-        channel.setMessageHandler((Object message) async {
+        channel.setMessageHandler((Object? message) async {
           // ignore message
           final WidgetUpdateResponse output = api.onWidgetUpdate();
           return output.encode();
@@ -490,12 +506,12 @@ abstract class FlutterTimerApi {
       }
     }
     {
-      const BasicMessageChannel<Object> channel =
-          BasicMessageChannel<Object>('dev.flutter.pigeon.FlutterTimerApi.onWidgetStartTimer', StandardMessageCodec());
+      const BasicMessageChannel<Object?> channel = BasicMessageChannel<Object?>(
+          'dev.flutter.pigeon.FlutterTimerApi.onWidgetStartTimer', StandardMessageCodec());
       if (api == null) {
         channel.setMessageHandler(null);
       } else {
-        channel.setMessageHandler((Object message) async {
+        channel.setMessageHandler((Object? message) async {
           // ignore message
           api.onWidgetStartTimer();
           return;
@@ -506,11 +522,19 @@ abstract class FlutterTimerApi {
 }
 
 class HostTimerApi {
+  /// Constructor for [HostTimerApi].  The [binaryMessenger] named argument is
+  /// available for dependency injection.  If it is left null, the default
+  /// BinaryMessenger will be used which routes to the host platform.
+  HostTimerApi({BinaryMessenger? binaryMessenger}) : _binaryMessenger = binaryMessenger;
+
+  final BinaryMessenger? _binaryMessenger;
+
   Future<void> init(InitializationRequest arg) async {
     final Object encoded = arg.encode();
-    const BasicMessageChannel<Object> channel =
-        BasicMessageChannel<Object>('dev.flutter.pigeon.HostTimerApi.init', StandardMessageCodec());
-    final Map<Object, Object> replyMap = await channel.send(encoded) as Map<Object, Object>;
+    final BasicMessageChannel<Object?> channel = BasicMessageChannel<Object?>(
+        'dev.flutter.pigeon.HostTimerApi.init', const StandardMessageCodec(), binaryMessenger: _binaryMessenger);
+    final Map<Object?, Object?>? replyMap =
+        await channel.send(encoded) as Map<Object?, Object?>?;
     if (replyMap == null) {
       throw PlatformException(
         code: 'channel-error',
@@ -518,10 +542,10 @@ class HostTimerApi {
         details: null,
       );
     } else if (replyMap['error'] != null) {
-      final Map<Object, Object> error = (replyMap['error'] as Map<Object, Object>);
+      final Map<Object?, Object?> error = (replyMap['error'] as Map<Object?, Object?>?)!;
       throw PlatformException(
-        code: (error['code'] as String),
-        message: error['message'] as String,
+        code: (error['code'] as String?)!,
+        message: error['message'] as String?,
         details: error['details'],
       );
     } else {
@@ -531,9 +555,10 @@ class HostTimerApi {
 
   Future<NotificationResponse> showRunningNotification(RunningNotificationRequest arg) async {
     final Object encoded = arg.encode();
-    const BasicMessageChannel<Object> channel =
-        BasicMessageChannel<Object>('dev.flutter.pigeon.HostTimerApi.showRunningNotification', StandardMessageCodec());
-    final Map<Object, Object> replyMap = await channel.send(encoded) as Map<Object, Object>;
+    final BasicMessageChannel<Object?> channel = BasicMessageChannel<Object?>(
+        'dev.flutter.pigeon.HostTimerApi.showRunningNotification', const StandardMessageCodec(), binaryMessenger: _binaryMessenger);
+    final Map<Object?, Object?>? replyMap =
+        await channel.send(encoded) as Map<Object?, Object?>?;
     if (replyMap == null) {
       throw PlatformException(
         code: 'channel-error',
@@ -541,22 +566,23 @@ class HostTimerApi {
         details: null,
       );
     } else if (replyMap['error'] != null) {
-      final Map<Object, Object> error = (replyMap['error'] as Map<Object, Object>);
+      final Map<Object?, Object?> error = (replyMap['error'] as Map<Object?, Object?>?)!;
       throw PlatformException(
-        code: (error['code'] as String),
-        message: error['message'] as String,
+        code: (error['code'] as String?)!,
+        message: error['message'] as String?,
         details: error['details'],
       );
     } else {
-      return NotificationResponse.decode(replyMap['result']);
+      return NotificationResponse.decode(replyMap['result']!);
     }
   }
 
   Future<NotificationResponse> showPausingNotification(TimeNotificationRequest arg) async {
     final Object encoded = arg.encode();
-    const BasicMessageChannel<Object> channel =
-        BasicMessageChannel<Object>('dev.flutter.pigeon.HostTimerApi.showPausingNotification', StandardMessageCodec());
-    final Map<Object, Object> replyMap = await channel.send(encoded) as Map<Object, Object>;
+    final BasicMessageChannel<Object?> channel = BasicMessageChannel<Object?>(
+        'dev.flutter.pigeon.HostTimerApi.showPausingNotification', const StandardMessageCodec(), binaryMessenger: _binaryMessenger);
+    final Map<Object?, Object?>? replyMap =
+        await channel.send(encoded) as Map<Object?, Object?>?;
     if (replyMap == null) {
       throw PlatformException(
         code: 'channel-error',
@@ -564,22 +590,23 @@ class HostTimerApi {
         details: null,
       );
     } else if (replyMap['error'] != null) {
-      final Map<Object, Object> error = (replyMap['error'] as Map<Object, Object>);
+      final Map<Object?, Object?> error = (replyMap['error'] as Map<Object?, Object?>?)!;
       throw PlatformException(
-        code: (error['code'] as String),
-        message: error['message'] as String,
+        code: (error['code'] as String?)!,
+        message: error['message'] as String?,
         details: error['details'],
       );
     } else {
-      return NotificationResponse.decode(replyMap['result']);
+      return NotificationResponse.decode(replyMap['result']!);
     }
   }
 
   Future<NotificationResponse> showElapsedNotification(NotificationRequest arg) async {
     final Object encoded = arg.encode();
-    const BasicMessageChannel<Object> channel =
-        BasicMessageChannel<Object>('dev.flutter.pigeon.HostTimerApi.showElapsedNotification', StandardMessageCodec());
-    final Map<Object, Object> replyMap = await channel.send(encoded) as Map<Object, Object>;
+    final BasicMessageChannel<Object?> channel = BasicMessageChannel<Object?>(
+        'dev.flutter.pigeon.HostTimerApi.showElapsedNotification', const StandardMessageCodec(), binaryMessenger: _binaryMessenger);
+    final Map<Object?, Object?>? replyMap =
+        await channel.send(encoded) as Map<Object?, Object?>?;
     if (replyMap == null) {
       throw PlatformException(
         code: 'channel-error',
@@ -587,22 +614,23 @@ class HostTimerApi {
         details: null,
       );
     } else if (replyMap['error'] != null) {
-      final Map<Object, Object> error = (replyMap['error'] as Map<Object, Object>);
+      final Map<Object?, Object?> error = (replyMap['error'] as Map<Object?, Object?>?)!;
       throw PlatformException(
-        code: (error['code'] as String),
-        message: error['message'] as String,
+        code: (error['code'] as String?)!,
+        message: error['message'] as String?,
         details: error['details'],
       );
     } else {
-      return NotificationResponse.decode(replyMap['result']);
+      return NotificationResponse.decode(replyMap['result']!);
     }
   }
 
   Future<CancelResponse> cancelTimer(CancelRequest arg) async {
     final Object encoded = arg.encode();
-    const BasicMessageChannel<Object> channel =
-        BasicMessageChannel<Object>('dev.flutter.pigeon.HostTimerApi.cancelTimer', StandardMessageCodec());
-    final Map<Object, Object> replyMap = await channel.send(encoded) as Map<Object, Object>;
+    final BasicMessageChannel<Object?> channel = BasicMessageChannel<Object?>(
+        'dev.flutter.pigeon.HostTimerApi.cancelTimer', const StandardMessageCodec(), binaryMessenger: _binaryMessenger);
+    final Map<Object?, Object?>? replyMap =
+        await channel.send(encoded) as Map<Object?, Object?>?;
     if (replyMap == null) {
       throw PlatformException(
         code: 'channel-error',
@@ -610,21 +638,46 @@ class HostTimerApi {
         details: null,
       );
     } else if (replyMap['error'] != null) {
-      final Map<Object, Object> error = (replyMap['error'] as Map<Object, Object>);
+      final Map<Object?, Object?> error = (replyMap['error'] as Map<Object?, Object?>?)!;
       throw PlatformException(
-        code: (error['code'] as String),
-        message: error['message'] as String,
+        code: (error['code'] as String?)!,
+        message: error['message'] as String?,
         details: error['details'],
       );
     } else {
-      return CancelResponse.decode(replyMap['result']);
+      return CancelResponse.decode(replyMap['result']!);
+    }
+  }
+
+  Future<void> toggleExtendByShake(ToggleRequest arg) async {
+    final Object encoded = arg.encode();
+    final BasicMessageChannel<Object?> channel = BasicMessageChannel<Object?>(
+        'dev.flutter.pigeon.HostTimerApi.toggleExtendByShake', const StandardMessageCodec(), binaryMessenger: _binaryMessenger);
+    final Map<Object?, Object?>? replyMap =
+        await channel.send(encoded) as Map<Object?, Object?>?;
+    if (replyMap == null) {
+      throw PlatformException(
+        code: 'channel-error',
+        message: 'Unable to establish connection on channel.',
+        details: null,
+      );
+    } else if (replyMap['error'] != null) {
+      final Map<Object?, Object?> error = (replyMap['error'] as Map<Object?, Object?>?)!;
+      throw PlatformException(
+        code: (error['code'] as String?)!,
+        message: error['message'] as String?,
+        details: error['details'],
+      );
+    } else {
+      // noop
     }
   }
 
   Future<InstalledAppsResponse> getInstalledPlayerApps() async {
-    const BasicMessageChannel<Object> channel =
-        BasicMessageChannel<Object>('dev.flutter.pigeon.HostTimerApi.getInstalledPlayerApps', StandardMessageCodec());
-    final Map<Object, Object> replyMap = await channel.send(null) as Map<Object, Object>;
+    final BasicMessageChannel<Object?> channel = BasicMessageChannel<Object?>(
+        'dev.flutter.pigeon.HostTimerApi.getInstalledPlayerApps', const StandardMessageCodec(), binaryMessenger: _binaryMessenger);
+    final Map<Object?, Object?>? replyMap =
+        await channel.send(null) as Map<Object?, Object?>?;
     if (replyMap == null) {
       throw PlatformException(
         code: 'channel-error',
@@ -632,21 +685,22 @@ class HostTimerApi {
         details: null,
       );
     } else if (replyMap['error'] != null) {
-      final Map<Object, Object> error = (replyMap['error'] as Map<Object, Object>);
+      final Map<Object?, Object?> error = (replyMap['error'] as Map<Object?, Object?>?)!;
       throw PlatformException(
-        code: (error['code'] as String),
-        message: error['message'] as String,
+        code: (error['code'] as String?)!,
+        message: error['message'] as String?,
         details: error['details'],
       );
     } else {
-      return InstalledAppsResponse.decode(replyMap['result']);
+      return InstalledAppsResponse.decode(replyMap['result']!);
     }
   }
 
   Future<InstalledAppsResponse> getInstalledAlarmApps() async {
-    const BasicMessageChannel<Object> channel =
-        BasicMessageChannel<Object>('dev.flutter.pigeon.HostTimerApi.getInstalledAlarmApps', StandardMessageCodec());
-    final Map<Object, Object> replyMap = await channel.send(null) as Map<Object, Object>;
+    final BasicMessageChannel<Object?> channel = BasicMessageChannel<Object?>(
+        'dev.flutter.pigeon.HostTimerApi.getInstalledAlarmApps', const StandardMessageCodec(), binaryMessenger: _binaryMessenger);
+    final Map<Object?, Object?>? replyMap =
+        await channel.send(null) as Map<Object?, Object?>?;
     if (replyMap == null) {
       throw PlatformException(
         code: 'channel-error',
@@ -654,22 +708,23 @@ class HostTimerApi {
         details: null,
       );
     } else if (replyMap['error'] != null) {
-      final Map<Object, Object> error = (replyMap['error'] as Map<Object, Object>);
+      final Map<Object?, Object?> error = (replyMap['error'] as Map<Object?, Object?>?)!;
       throw PlatformException(
-        code: (error['code'] as String),
-        message: error['message'] as String,
+        code: (error['code'] as String?)!,
+        message: error['message'] as String?,
         details: error['details'],
       );
     } else {
-      return InstalledAppsResponse.decode(replyMap['result']);
+      return InstalledAppsResponse.decode(replyMap['result']!);
     }
   }
 
   Future<void> launchApp(LaunchAppRequest arg) async {
     final Object encoded = arg.encode();
-    const BasicMessageChannel<Object> channel =
-        BasicMessageChannel<Object>('dev.flutter.pigeon.HostTimerApi.launchApp', StandardMessageCodec());
-    final Map<Object, Object> replyMap = await channel.send(encoded) as Map<Object, Object>;
+    final BasicMessageChannel<Object?> channel = BasicMessageChannel<Object?>(
+        'dev.flutter.pigeon.HostTimerApi.launchApp', const StandardMessageCodec(), binaryMessenger: _binaryMessenger);
+    final Map<Object?, Object?>? replyMap =
+        await channel.send(encoded) as Map<Object?, Object?>?;
     if (replyMap == null) {
       throw PlatformException(
         code: 'channel-error',
@@ -677,10 +732,10 @@ class HostTimerApi {
         details: null,
       );
     } else if (replyMap['error'] != null) {
-      final Map<Object, Object> error = (replyMap['error'] as Map<Object, Object>);
+      final Map<Object?, Object?> error = (replyMap['error'] as Map<Object?, Object?>?)!;
       throw PlatformException(
-        code: (error['code'] as String),
-        message: error['message'] as String,
+        code: (error['code'] as String?)!,
+        message: error['message'] as String?,
         details: error['details'],
       );
     } else {
@@ -689,9 +744,10 @@ class HostTimerApi {
   }
 
   Future<Package> dummyApp() async {
-    const BasicMessageChannel<Object> channel =
-        BasicMessageChannel<Object>('dev.flutter.pigeon.HostTimerApi.dummyApp', StandardMessageCodec());
-    final Map<Object, Object> replyMap = await channel.send(null) as Map<Object, Object>;
+    final BasicMessageChannel<Object?> channel = BasicMessageChannel<Object?>(
+        'dev.flutter.pigeon.HostTimerApi.dummyApp', const StandardMessageCodec(), binaryMessenger: _binaryMessenger);
+    final Map<Object?, Object?>? replyMap =
+        await channel.send(null) as Map<Object?, Object?>?;
     if (replyMap == null) {
       throw PlatformException(
         code: 'channel-error',
@@ -699,14 +755,14 @@ class HostTimerApi {
         details: null,
       );
     } else if (replyMap['error'] != null) {
-      final Map<Object, Object> error = (replyMap['error'] as Map<Object, Object>);
+      final Map<Object?, Object?> error = (replyMap['error'] as Map<Object?, Object?>?)!;
       throw PlatformException(
-        code: (error['code'] as String),
-        message: error['message'] as String,
+        code: (error['code'] as String?)!,
+        message: error['message'] as String?,
         details: error['details'],
       );
     } else {
-      return Package.decode(replyMap['result']);
+      return Package.decode(replyMap['result']!);
     }
   }
 }
