@@ -1,4 +1,3 @@
-import 'package:flutter/foundation.dart';
 import 'package:in_app_purchase/in_app_purchase.dart';
 
 class Product {
@@ -7,7 +6,7 @@ class Product {
   final bool purchased;
 
   Product(
-      {@required this.productDetails,
-      @required this.consumable,
+      {required this.productDetails,
+      required this.consumable,
       this.purchased = false});
 }
