@@ -1,5 +1,5 @@
 import 'package:shared_preferences/shared_preferences.dart';
-import 'package:sleep_timer/app/auto_router.gr.dart';
+import 'package:sleep_timer/app/app.router.dart';
 import 'package:sleep_timer/app/locator.dart';
 import 'package:sleep_timer/common/constants.dart';
 import 'package:sleep_timer/common/timer_service_manager.dart';
@@ -11,37 +11,37 @@ import 'package:stacked/stacked.dart';
 import 'package:stacked_services/stacked_services.dart';
 
 class HomeViewModel extends BaseViewModel {
-  final _navigationService = locator<NavigationService>();
-  final _prefService = locator<SharedPreferences>();
-  final _reviewService = locator<ReviewService>();
-  final _themeService = locator<ThemeService>();
+  final NavigationService? _navigationService = locator<NavigationService>();
+  final SharedPreferences? _prefService = locator<SharedPreferences>();
+  final ReviewService? _reviewService = locator<ReviewService>();
+  final ThemeService? _themeService = locator<ThemeService>();
 
-  bool get showGlow => _themeService.showGlow;
+  bool get showGlow => _themeService!.showGlow;
 
   int _initialTime = kDefaultInitialTime;
   int get initialTime => _initialTime;
 
-  String activeTimerId;
+  String? activeTimerId;
   bool get hasActiveTimer => activeTimerId != null;
 
   // lock variable to prevent multiple calls due to rebuild
   bool mayAskForReviewLocked = false;
 
   HomeViewModel() {
-    _initialTime = _prefService.getInt(kPrefKeyInitialTime) ?? _initialTime;
+    _initialTime = _prefService!.getInt(kPrefKeyInitialTime) ?? _initialTime;
 
     startActionList.forEach((element) {
       element.enabled =
-          _prefService.getBool(element.id.toString()) ?? element.enabled;
+          _prefService!.getBool(element.id.toString()) ?? element.enabled;
 
       if (element is ValueActionModel) {
-        element.value = _prefService.getDouble(element.key) ?? element.value;
+        element.value = _prefService!.getDouble(element.key!) ?? element.value;
       }
     });
 
     endActionList.forEach((element) {
       element.enabled =
-          _prefService.get(element.id.toString()) ?? element.enabled;
+          _prefService!.get(element.id.toString()) as bool? ?? element.enabled;
     });
   }
 
@@ -49,19 +49,19 @@ class HomeViewModel extends BaseViewModel {
     final timerModel =
         TimerModel(_initialTime * 60, startActionList, endActionList);
 
-    activeTimerId = await _navigationService.navigateTo(Routes.timerView,
-        arguments: TimerViewArguments(timerModel: timerModel));
+    activeTimerId = await (_navigationService!.navigateTo(Routes.timerView,
+        arguments: TimerViewArguments(timerModel: timerModel)));
     notifyListeners();
   }
 
   void openActiveTimer() async {
     if (hasActiveTimer) {
-      final timerModel = TimerServiceManager.getInstance()
-          .getTimerService(activeTimerId)
+      final timerModel = TimerServiceManager.instance
+          .getTimerService(activeTimerId)!
           .timerModel;
 
-      activeTimerId = await _navigationService.navigateTo(Routes.timerView,
-          arguments: TimerViewArguments(timerModel: timerModel));
+      activeTimerId = await (_navigationService!.navigateTo(Routes.timerView,
+          arguments: TimerViewArguments(timerModel: timerModel)));
       notifyListeners();
     }
   }
@@ -69,7 +69,7 @@ class HomeViewModel extends BaseViewModel {
   /// Sets the time without notifying listeners.
   void setTimeSilent(int value) {
     _initialTime = value;
-    _prefService.setInt(kPrefKeyInitialTime, value);
+    _prefService!.setInt(kPrefKeyInitialTime, value);
   }
 
   void setTime(int value) {
@@ -81,14 +81,13 @@ class HomeViewModel extends BaseViewModel {
     if (!mayAskForReviewLocked && !hasActiveTimer) {
       // to prevent multiple calls
       mayAskForReviewLocked = true;
-      final shouldAsk = _reviewService.shouldAskForReview();
+      final shouldAsk = _reviewService!.shouldAskForReview();
       if (shouldAsk) {
-        await _reviewService.requestReview();
+        await _reviewService!.requestReview();
       }
     }
   }
 
-  Future navigateToSettings() async {
-    await _navigationService.navigateTo(Routes.settingsView);
-  }
+  Future navigateToSettings() async =>
+      await _navigationService!.navigateTo(Routes.settingsView);
 }
