@@ -38,74 +38,49 @@ class CircularSliderAppearance {
   final bool counterClockwise;
   final double animDurationMultiplier;
   final int spinnerDuration;
-  final CustomSliderWidths customWidths;
-  final ValueNotifier<CustomSliderColors> customColors;
-  final InfoProperties infoProperties;
+  final CustomSliderWidths? customWidths;
+  final ValueNotifier<CustomSliderColors?>? customColors;
+  final InfoProperties? infoProperties;
 
-  double get _customTrackWidth =>
-      customWidths != null ? customWidths.trackWidth : null;
-  double get _customProgressBarWidth =>
-      customWidths != null ? customWidths.progressBarWidth : null;
-  double get _customShadowWidth =>
-      customWidths != null ? customWidths.shadowWidth : null;
-  double get _customHandlerSize =>
-      customWidths != null ? customWidths.handlerSize : null;
+  double? get _customTrackWidth => customWidths?.trackWidth;
+  double? get _customProgressBarWidth => customWidths?.progressBarWidth;
+  double? get _customShadowWidth => customWidths?.shadowWidth;
+  double? get _customHandlerSize => customWidths?.handlerSize;
 
   double get trackWidth => _customTrackWidth ?? progressBarWidth / 4.0;
   double get progressBarWidth => _customProgressBarWidth ?? size / 10.0;
   double get handlerSize => _customHandlerSize ?? progressBarWidth / 5.0;
   double get shadowWidth => _customShadowWidth ?? progressBarWidth * 1.4;
 
-  Color get _customTrackColor =>
-      customColors != null ? customColors.value.trackColor : null;
-  List<Color> get _customProgressBarColors {
-    if (customColors != null) {
-      if (customColors.value.progressBarColors != null) {
-        return customColors.value.progressBarColors;
-      } else if (customColors.value.progressBarColor != null) {
-        return [
-          customColors.value.progressBarColor,
-          customColors.value.progressBarColor
-        ];
+  Color? get _customTrackColor => customColors?.value?.trackColor;
+  List<Color>? get _customProgressBarColors {
+    if (customColors?.value != null) {
+      if (customColors!.value!.progressBarColors != null) {
+        return customColors!.value!.progressBarColors;
+      } else if (customColors!.value!.progressBarColor != null) {
+        return [customColors!.value!.progressBarColor!, customColors!.value!.progressBarColor!];
       }
     }
     return null;
   }
 
-  List<Color> get _customTrackColors {
-    if (customColors != null) {
-      if (customColors.value.trackColors != null) {
-        return customColors.value.trackColors;
-      } else {
-        return null;
-      }
-    }
-    return null;
+  List<Color>? get _customTrackColors {
+  	return customColors?.value?.trackColors;
   }
 
-  double get _gradientStartAngle =>
-      customColors != null ? customColors.value.gradientStartAngle : null;
-  double get _gradientEndAngle =>
-      customColors != null ? customColors.value.gradientEndAngle : null;
-  double get _trackGradientStartAngle =>
-      customColors != null ? customColors.value.trackGradientStartAngle : null;
-  double get _trackGradientEndAngle =>
-      customColors != null ? customColors.value.trackGradientEndAngle : null;
-  bool get _dynamicGradient =>
-      customColors != null ? customColors.value.dynamicGradient : null;
-  Color get _customShadowColor =>
-      customColors != null ? customColors.value.shadowColor : null;
-  double get _customShadowMaxOpacity =>
-      customColors != null ? customColors.value.shadowMaxOpacity : null;
-  double get _customShadowStep =>
-      customColors != null ? customColors.value.shadowStep : null;
-  Color get _customDotColor =>
-      customColors != null ? customColors.value.dotColor : null;
-  bool get _hideShadow =>
-      customColors != null ? customColors.value.hideShadow : null;
+  double? get _gradientStartAngle => customColors?.value?.gradientStartAngle;
+  double? get _gradientEndAngle => customColors?.value?.gradientEndAngle;
+  double? get _trackGradientStartAngle => customColors?.value?.trackGradientStartAngle;
+  double? get _trackGradientEndAngle => customColors?.value?.trackGradientEndAngle;
+  bool? get _dynamicGradient => customColors?.value?.dynamicGradient;
+  Color? get _customShadowColor => customColors?.value?.shadowColor;
+  double? get _customShadowMaxOpacity => customColors?.value?.shadowMaxOpacity;
+  double? get _customShadowStep => customColors?.value?.shadowStep;
+  Color? get _customDotColor => customColors?.value?.dotColor;
+  bool? get _hideShadow => customColors?.value?.hideShadow;
 
   Color get trackColor => _customTrackColor ?? _defaultTrackColor;
-  List<Color> get trackColors => _customTrackColors;
+  List<Color>? get trackColors => _customTrackColors;
   List<Color> get progressBarColors =>
       _customProgressBarColors ?? _defaultBarColors;
   double get gradientStartAngle =>
@@ -120,46 +95,36 @@ class CircularSliderAppearance {
   Color get shadowColor => _customShadowColor ?? _defaultShadowColor;
   double get shadowMaxOpacity =>
       _customShadowMaxOpacity ?? _defaultShadowMaxOpacity;
-  double get shadowStep => _customShadowStep;
+  double? get shadowStep => _customShadowStep;
   Color get dotColor => _customDotColor ?? _defaultDotColor;
 
-  String get _topLabelText =>
-      infoProperties != null ? infoProperties.topLabelText : null;
-  String get _bottomLabelText =>
-      infoProperties != null ? infoProperties.bottomLabelText : null;
-  TextStyle get _mainLabelStyle =>
-      infoProperties != null ? infoProperties.mainLabelStyle : null;
-  TextStyle get _topLabelStyle =>
-      infoProperties != null ? infoProperties.topLabelStyle : null;
-  TextStyle get _bottomLabelStyle =>
-      infoProperties != null ? infoProperties.bottomLabelStyle : null;
-  PercentageModifier get _modifier =>
-      infoProperties != null ? infoProperties.modifier : null;
+  String? get _topLabelText => infoProperties?.topLabelText;
+  String? get _bottomLabelText => infoProperties?.bottomLabelText;
+  TextStyle? get _mainLabelStyle => infoProperties?.mainLabelStyle;
+  TextStyle? get _topLabelStyle => infoProperties?.topLabelStyle;
+  TextStyle? get _bottomLabelStyle => infoProperties?.bottomLabelStyle;
+  PercentageModifier? get _modifier => infoProperties?.modifier;
 
   PercentageModifier get infoModifier =>
       _modifier ?? _defaultPercentageModifier;
-  String get infoTopLabelText => _topLabelText;
-  String get infoBottomLabelText => _bottomLabelText;
+  String? get infoTopLabelText => _topLabelText;
+  String? get infoBottomLabelText => _bottomLabelText;
   TextStyle get infoMainLabelStyle {
-    if (_mainLabelStyle != null) return _mainLabelStyle;
-
-    return TextStyle(
+    return _mainLabelStyle ?? TextStyle(
         fontWeight: FontWeight.w100,
         fontSize: size / 5.0,
         color: Color.fromRGBO(30, 0, 59, 1.0));
   }
 
   TextStyle get infoTopLabelStyle {
-    if (_topLabelStyle != null) return _topLabelStyle;
-    return TextStyle(
+    return _topLabelStyle ?? TextStyle(
         fontWeight: FontWeight.w600,
         fontSize: size / 10.0,
         color: Color.fromRGBO(147, 81, 120, 1.0));
   }
 
   TextStyle get infoBottomLabelStyle {
-    if (_bottomLabelStyle != null) return _bottomLabelStyle;
-    return TextStyle(
+    return _bottomLabelStyle ?? TextStyle(
         fontWeight: FontWeight.w600,
         fontSize: size / 10.0,
         color: Color.fromRGBO(147, 81, 120, 1.0));
@@ -180,10 +145,10 @@ class CircularSliderAppearance {
 }
 
 class CustomSliderWidths {
-  final double trackWidth;
-  final double progressBarWidth;
-  final double handlerSize;
-  final double shadowWidth;
+  final double? trackWidth;
+  final double? progressBarWidth;
+  final double? handlerSize;
+  final double? shadowWidth;
 
   CustomSliderWidths(
       {this.trackWidth,
@@ -193,20 +158,20 @@ class CustomSliderWidths {
 }
 
 class CustomSliderColors {
-  final Color trackColor;
-  final Color progressBarColor;
-  final List<Color> progressBarColors;
-  final double gradientStartAngle;
-  final double gradientEndAngle;
-  final List<Color> trackColors;
-  final double trackGradientStartAngle;
-  final double trackGradientEndAngle;
+  final Color? trackColor;
+  final Color? progressBarColor;
+  final List<Color>? progressBarColors;
+  final double? gradientStartAngle;
+  final double? gradientEndAngle;
+  final List<Color>? trackColors;
+  final double? trackGradientStartAngle;
+  final double? trackGradientEndAngle;
   final bool dynamicGradient;
-  final bool hideShadow;
-  final Color shadowColor;
-  final double shadowMaxOpacity;
-  final double shadowStep;
-  final Color dotColor;
+  final bool? hideShadow;
+  final Color? shadowColor;
+  final double? shadowMaxOpacity;
+  final double? shadowStep;
+  final Color? dotColor;
 
   CustomSliderColors(
       {this.trackColor,
@@ -226,12 +191,12 @@ class CustomSliderColors {
 }
 
 class InfoProperties {
-  final PercentageModifier modifier;
-  final TextStyle mainLabelStyle;
-  final TextStyle topLabelStyle;
-  final TextStyle bottomLabelStyle;
-  final String topLabelText;
-  final String bottomLabelText;
+  final PercentageModifier? modifier;
+  final TextStyle? mainLabelStyle;
+  final TextStyle? topLabelStyle;
+  final TextStyle? bottomLabelStyle;
+  final String? topLabelText;
+  final String? bottomLabelText;
 
   InfoProperties(
       {this.topLabelText,
