@@ -22,9 +22,9 @@ class SleekCircularSlider extends StatefulWidget {
   final int minValue;
   final int maxValue;
   final CircularSliderAppearance appearance;
-  final OnChange onChange;
-  final OnChange onChangeEnd;
-  final InnerWidget innerWidget;
+  final OnChange? onChange;
+  final OnChange? onChangeEnd;
+  final InnerWidget? innerWidget;
   static const defaultAppearance = CircularSliderAppearance();
 
   double get angle {
@@ -33,7 +33,7 @@ class SleekCircularSlider extends StatefulWidget {
   }
 
   const SleekCircularSlider(
-      {Key key,
+      {Key? key,
       this.initialValue = 50,
       this.minValue = 0,
       this.maxValue = 100,
@@ -41,12 +41,8 @@ class SleekCircularSlider extends StatefulWidget {
       this.onChange,
       this.onChangeEnd,
       this.innerWidget})
-      : assert(initialValue != null),
-        assert(minValue != null),
-        assert(maxValue != null),
-        assert(minValue <= maxValue),
+      : assert(minValue <= maxValue),
         assert(initialValue >= minValue && initialValue <= maxValue),
-        assert(appearance != null),
         super(key: key);
   @override
   _SleekCircularSliderState createState() => _SleekCircularSliderState();
@@ -56,14 +52,14 @@ class _SleekCircularSliderState extends State<SleekCircularSlider>
     with SingleTickerProviderStateMixin {
   bool _isHandlerSelected = false;
   bool _animationInProgress = false;
-  _CurvePainter _painter;
-  double _oldWidgetAngle;
-  double _oldWidgetValue;
-  double _currentAngle;
-  double _startAngle;
-  double _angleRange;
-  double _selectedAngle;
-  ValueChangedAnimationManager _animationManager;
+  _CurvePainter? _painter;
+  double? _oldWidgetAngle;
+  double? _oldWidgetValue;
+  double? _currentAngle;
+  late double _startAngle;
+  late double _angleRange;
+  double? _selectedAngle;
+  ValueChangedAnimationManager? _animationManager;
 
   bool get _interactionEnabled =>
       (widget.onChangeEnd != null || widget.onChange != null);
@@ -98,7 +94,7 @@ class _SleekCircularSliderState extends State<SleekCircularSlider>
       durationMultiplier: widget.appearance.animDurationMultiplier,
     );
 
-    _animationManager.animate(
+    _animationManager?.animate(
         initialValue: widget.initialValue.toDouble(),
         angle: widget.angle,
         oldAngle: _oldWidgetAngle,
@@ -140,7 +136,7 @@ class _SleekCircularSliderState extends State<SleekCircularSlider>
 
   @override
   void dispose() {
-    if (_animationManager != null) _animationManager.dispose();
+    _animationManager?.dispose();
     super.dispose();
   }
 
@@ -163,7 +159,7 @@ class _SleekCircularSliderState extends State<SleekCircularSlider>
     _painter = _CurvePainter(
         startAngle: _startAngle,
         angleRange: _angleRange,
-        angle: _currentAngle < 0.5 ? 0.5 : _currentAngle,
+        angle: _currentAngle! < 0.5 ? 0.5 : _currentAngle!,
         appearance: widget.appearance);
     _oldWidgetAngle = widget.angle;
     _oldWidgetValue = widget.initialValue.toDouble();
@@ -172,16 +168,16 @@ class _SleekCircularSliderState extends State<SleekCircularSlider>
   void _updateOnChange() {
     if (widget.onChange != null && !_animationInProgress) {
       final value = angleToValue(
-          _currentAngle, widget.minValue, widget.maxValue, _angleRange);
-      widget.onChange(value.round());
+          _currentAngle!, widget.minValue, widget.maxValue, _angleRange);
+      widget.onChange!(value.round());
     }
   }
 
   Widget _buildChildWidget() {
     final value = angleToValue(
-        _currentAngle, widget.minValue, widget.maxValue, _angleRange);
+        _currentAngle!, widget.minValue, widget.maxValue, _angleRange);
     final childWidget = widget.innerWidget != null
-        ? widget.innerWidget(value)
+        ? widget.innerWidget!(value)
         : SliderLabel(value: value, appearance: widget.appearance);
     return childWidget;
   }
@@ -190,7 +186,7 @@ class _SleekCircularSliderState extends State<SleekCircularSlider>
     if (!_isHandlerSelected) {
       return;
     }
-    if (_painter.center == null) {
+    if (_painter?.center == null) {
       return;
     }
     _handlePan(details, false);
@@ -200,22 +196,22 @@ class _SleekCircularSliderState extends State<SleekCircularSlider>
     _handlePan(details, true);
     if (widget.onChangeEnd != null) {
       final value = angleToValue(
-          _currentAngle, widget.minValue, widget.maxValue, _angleRange);
-      widget.onChangeEnd(value.round());
+          _currentAngle!, widget.minValue, widget.maxValue, _angleRange);
+      widget.onChangeEnd!(value.round());
     }
   }
 
   void _handlePan(Offset details, bool isPanEnd) {
-    if (_painter.center == null) {
+    if (_painter?.center == null) {
       return;
     }
-    RenderBox renderBox = context.findRenderObject();
+    var renderBox = context.findRenderObject() as RenderBox;
     var position = renderBox.globalToLocal(details);
     final touchWidth = widget.appearance.progressBarWidth + 100;
 
     if (isPointAlongCircle(
-        position, _painter.center, _painter.radius, touchWidth)) {
-      _selectedAngle = coordinatesToRadians(_painter.center, position);
+        position, _painter!.center!, _painter!.radius, touchWidth)) {
+      _selectedAngle = coordinatesToRadians(_painter!.center!, position);
       // setup painter with new angle values and update onChange
       _setupPainter(counterClockwise: widget.appearance.counterClockwise);
       _updateOnChange();
@@ -224,20 +220,18 @@ class _SleekCircularSliderState extends State<SleekCircularSlider>
   }
 
   bool _onPanDown(Offset details) {
-    if (_painter == null || _interactionEnabled == false) {
+    if (_painter == null ||
+        _painter?.center == null ||
+        _interactionEnabled == false) {
       return false;
     }
-    RenderBox renderBox = context.findRenderObject();
+    var renderBox = context.findRenderObject() as RenderBox;
     var position = renderBox.globalToLocal(details);
-
-    if (position == null) {
-      return false;
-    }
 
     final angleWithinRange = isAngleWithinRange(
         startAngle: _startAngle,
         angleRange: _angleRange,
-        touchAngle: coordinatesToRadians(_painter.center, position),
+        touchAngle: coordinatesToRadians(_painter!.center!, position),
         previousAngle: _currentAngle,
         counterClockwise: widget.appearance.counterClockwise);
     if (!angleWithinRange) {
@@ -247,7 +241,7 @@ class _SleekCircularSliderState extends State<SleekCircularSlider>
     final touchWidth = widget.appearance.handlerSize + 10;
 
     if (isPointAlongCircle(
-        position, _painter.center, _painter.radius, touchWidth)) {
+        position, _painter!.center!, _painter!.radius, touchWidth)) {
       _isHandlerSelected = true;
       _onPanUpdate(details);
     } else {
