@@ -3,21 +3,19 @@ part of circular_slider;
 class _CurvePainter extends CustomPainter {
   final double angle;
   final CircularSliderAppearance appearance;
-  final startAngle;
-  final angleRange;
+  final double startAngle;
+  final double angleRange;
 
-  Offset handler;
-  Offset center;
-  double radius;
+  Offset? handler;
+  Offset? center;
+  late double radius;
 
-  _CurvePainter(
-      {@required this.appearance,
-      this.angle = 30,
-      @required this.startAngle,
-      @required this.angleRange})
-      : assert(appearance != null),
-        assert(startAngle != null),
-        assert(angleRange != null);
+  _CurvePainter({
+    required this.appearance,
+    this.angle = 30,
+    required this.startAngle,
+    required this.angleRange,
+  });
 
   @override
   void paint(Canvas canvas, Size size) {
@@ -34,7 +32,7 @@ class _CurvePainter extends CustomPainter {
         startAngle: degreeToRadians(appearance.trackGradientStartAngle),
         endAngle: degreeToRadians(appearance.trackGradientStopAngle),
         tileMode: TileMode.mirror,
-        colors: appearance.trackColors,
+        colors: appearance.trackColors!,
       );
       trackPaint = Paint()
         ..shader = trackGradient.createShader(progressBarRect)
@@ -60,7 +58,7 @@ class _CurvePainter extends CustomPainter {
     }
 
     final currentAngle = appearance.counterClockwise ? -angle : angle;
-    final dynamicGradient = appearance.dynamicGradient ?? false;
+    final dynamicGradient = appearance.dynamicGradient;
     final gradientRotationAngle = dynamicGradient
         ? appearance.counterClockwise
             ? startAngle + 10.0
@@ -105,31 +103,30 @@ class _CurvePainter extends CustomPainter {
     var dotPaint = Paint()..color = appearance.dotColor;
 
     var handler = degreesToCoordinates(
-        center, -math.pi / 2 + startAngle + currentAngle + 1.5, radius);
+        center!, -math.pi / 2 + startAngle + currentAngle + 1.5, radius);
     canvas.drawCircle(handler, appearance.handlerSize, dotPaint);
   }
 
   void drawCircularArc(
-      {@required Canvas canvas,
-      @required Size size,
-      @required Paint paint,
+      {required Canvas canvas,
+      required Size size,
+      required Paint paint,
       bool ignoreAngle = false,
       bool spinnerMode = false}) {
-    final double angleValue = ignoreAngle ? 0 : (angleRange - angle);
+    final angleValue = ignoreAngle ? 0 : (angleRange - angle);
     final range = appearance.counterClockwise ? -angleRange : angleRange;
     final currentAngle = appearance.counterClockwise ? angleValue : -angleValue;
     canvas.drawArc(
-        Rect.fromCircle(center: center, radius: radius),
+        Rect.fromCircle(center: center!, radius: radius),
         degreeToRadians(spinnerMode ? 0 : startAngle),
         degreeToRadians(spinnerMode ? 360 : range + currentAngle),
         false,
         paint);
   }
 
-  void drawShadow({@required Canvas canvas, @required Size size}) {
-    final int shadowStep = appearance.shadowStep ??
-        math.max(
-            1, (appearance.shadowWidth - appearance.progressBarWidth) ~/ 10);
+  void drawShadow({required Canvas canvas, required Size size}) {
+    final shadowStep = appearance.shadowStep ?? math.max(
+            1, (appearance.shadowWidth - appearance.progressBarWidth) ~/ 10).toDouble();
     final maxOpacity = math.min(1.0, appearance.shadowMaxOpacity);
     final repetitions = math.max(1,
         ((appearance.shadowWidth - appearance.progressBarWidth) ~/ shadowStep));
