@@ -4,7 +4,12 @@ import 'appearance.dart';
 class SliderLabel extends StatelessWidget {
   final double value;
   final CircularSliderAppearance appearance;
-  const SliderLabel({Key key, this.value, this.appearance}) : super(key: key);
+
+  const SliderLabel({
+    Key? key,
+    required this.value,
+    required this.appearance,
+  }) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
@@ -18,7 +23,7 @@ class SliderLabel extends StatelessWidget {
     var widgets = <Widget>[];
     if (appearance.infoTopLabelText != null) {
       widgets.add(Text(
-        appearance.infoTopLabelText,
+        appearance.infoTopLabelText!,
         style: appearance.infoTopLabelStyle,
       ));
     }
@@ -28,7 +33,7 @@ class SliderLabel extends StatelessWidget {
     );
     if (appearance.infoBottomLabelText != null) {
       widgets.add(Text(
-        appearance.infoBottomLabelText,
+        appearance.infoBottomLabelText!,
         style: appearance.infoBottomLabelStyle,
       ));
     }
