@@ -1,4 +1,4 @@
-// Autogenerated from Pigeon (v0.1.23), do not edit directly.
+// Autogenerated from Pigeon (v0.2.1), do not edit directly.
 // See also: https://pub.dev/packages/pigeon
 
 package dr.achim.sleep_timer;
@@ -15,7 +15,7 @@ import java.util.HashMap;
 public class Messages {
 
   /** Generated class from Pigeon that represents data sent in messages. */
-  public static class ExtendTimeResponse {
+  public static class ExtendTimeRequest {
     private String timerId;
     public String getTimerId() { return timerId; }
     public void setTimerId(String setterArg) { this.timerId = setterArg; }
@@ -30,8 +30,8 @@ public class Messages {
       toMapResult.put("additionalTime", additionalTime);
       return toMapResult;
     }
-    static ExtendTimeResponse fromMap(Map<String, Object> map) {
-      ExtendTimeResponse fromMapResult = new ExtendTimeResponse();
+    static ExtendTimeRequest fromMap(Map<String, Object> map) {
+      ExtendTimeRequest fromMapResult = new ExtendTimeRequest();
       Object timerId = map.get("timerId");
       fromMapResult.timerId = (String)timerId;
       Object additionalTime = map.get("additionalTime");
@@ -474,6 +474,25 @@ public class Messages {
   }
 
   /** Generated class from Pigeon that represents data sent in messages. */
+  public static class ToggleRequest {
+    private Boolean enable;
+    public Boolean getEnable() { return enable; }
+    public void setEnable(Boolean setterArg) { this.enable = setterArg; }
+
+    Map<String, Object> toMap() {
+      Map<String, Object> toMapResult = new HashMap<>();
+      toMapResult.put("enable", enable);
+      return toMapResult;
+    }
+    static ToggleRequest fromMap(Map<String, Object> map) {
+      ToggleRequest fromMapResult = new ToggleRequest();
+      Object enable = map.get("enable");
+      fromMapResult.enable = (Boolean)enable;
+      return fromMapResult;
+    }
+  }
+
+  /** Generated class from Pigeon that represents data sent in messages. */
   public static class InstalledAppsResponse {
     private List<Object> apps;
     public List<Object> getApps() { return apps; }
@@ -553,7 +572,7 @@ public class Messages {
     public interface Reply<T> {
       void reply(T reply);
     }
-    public void onExtendTime(ExtendTimeResponse argInput, Reply<Void> callback) {
+    public void onExtendTime(ExtendTimeRequest argInput, Reply<Void> callback) {
       BasicMessageChannel<Object> channel =
           new BasicMessageChannel<>(binaryMessenger, "dev.flutter.pigeon.FlutterTimerApi.onExtendTime", new StandardMessageCodec());
       Map<String, Object> inputMap = argInput.toMap();
@@ -643,6 +662,7 @@ public class Messages {
     NotificationResponse showPausingNotification(TimeNotificationRequest arg);
     NotificationResponse showElapsedNotification(NotificationRequest arg);
     CancelResponse cancelTimer(CancelRequest arg);
+    void toggleExtendByShake(ToggleRequest arg);
     InstalledAppsResponse getInstalledPlayerApps();
     InstalledAppsResponse getInstalledAlarmApps();
     void launchApp(LaunchAppRequest arg);
@@ -745,6 +765,27 @@ public class Messages {
               CancelRequest input = CancelRequest.fromMap((Map<String, Object>)message);
               CancelResponse output = api.cancelTimer(input);
               wrapped.put("result", output.toMap());
+            }
+            catch (Error | RuntimeException exception) {
+              wrapped.put("error", wrapError(exception));
+            }
+            reply.reply(wrapped);
+          });
+        } else {
+          channel.setMessageHandler(null);
+        }
+      }
+      {
+        BasicMessageChannel<Object> channel =
+            new BasicMessageChannel<>(binaryMessenger, "dev.flutter.pigeon.HostTimerApi.toggleExtendByShake", new StandardMessageCodec());
+        if (api != null) {
+          channel.setMessageHandler((message, reply) -> {
+            Map<String, Object> wrapped = new HashMap<>();
+            try {
+              @SuppressWarnings("ConstantConditions")
+              ToggleRequest input = ToggleRequest.fromMap((Map<String, Object>)message);
+              api.toggleExtendByShake(input);
+              wrapped.put("result", null);
             }
             catch (Error | RuntimeException exception) {
               wrapped.put("error", wrapError(exception));
