@@ -8,16 +8,15 @@ part of 'bridge_model.dart';
 
 BridgeModel _$BridgeModelFromJson(Map<String, dynamic> json) {
   return BridgeModel(
-    id: json['id'] as String,
-    ip: json['ip'] as String,
-    mac: json['mac'] as String,
-    name: json['name'] as String,
-    auth: json['auth'] as String,
+    id: json['id'] as String?,
+    ip: json['ip'] as String?,
+    mac: json['mac'] as String?,
+    name: json['name'] as String?,
+    auth: json['auth'] as String?,
     state: _$enumDecodeNullable(_$ConnectionEnumMap, json['state']),
-    groups: (json['groups'] as List)
-        ?.map((e) =>
-            e == null ? null : LightGroup.fromJson(e as Map<String, dynamic>))
-        ?.toList(),
+    groups: (json['groups'] as List<dynamic>?)
+        ?.map((e) => LightGroup.fromJson(e as Map<String, dynamic>))
+        .toList(),
   );
 }
 
@@ -32,36 +31,41 @@ Map<String, dynamic> _$BridgeModelToJson(BridgeModel instance) =>
       'groups': instance.groups,
     };
 
-T _$enumDecode<T>(
-  Map<T, dynamic> enumValues,
-  dynamic source, {
-  T unknownValue,
+K _$enumDecode<K, V>(
+  Map<K, V> enumValues,
+  Object? source, {
+  K? unknownValue,
 }) {
   if (source == null) {
-    throw ArgumentError('A value must be provided. Supported values: '
-        '${enumValues.values.join(', ')}');
+    throw ArgumentError(
+      'A value must be provided. Supported values: '
+      '${enumValues.values.join(', ')}',
+    );
   }
 
-  final value = enumValues.entries
-      .singleWhere((e) => e.value == source, orElse: () => null)
-      ?.key;
-
-  if (value == null && unknownValue == null) {
-    throw ArgumentError('`$source` is not one of the supported values: '
-        '${enumValues.values.join(', ')}');
-  }
-  return value ?? unknownValue;
+  return enumValues.entries.singleWhere(
+    (e) => e.value == source,
+    orElse: () {
+      if (unknownValue == null) {
+        throw ArgumentError(
+          '`$source` is not one of the supported values: '
+          '${enumValues.values.join(', ')}',
+        );
+      }
+      return MapEntry(unknownValue, enumValues.values.first);
+    },
+  ).key;
 }
 
-T _$enumDecodeNullable<T>(
-  Map<T, dynamic> enumValues,
+K? _$enumDecodeNullable<K, V>(
+  Map<K, V> enumValues,
   dynamic source, {
-  T unknownValue,
+  K? unknownValue,
 }) {
   if (source == null) {
     return null;
   }
-  return _$enumDecode<T>(enumValues, source, unknownValue: unknownValue);
+  return _$enumDecode<K, V>(enumValues, source, unknownValue: unknownValue);
 }
 
 const _$ConnectionEnumMap = {
