@@ -10,8 +10,8 @@ import 'package:sleep_timer/services/device_service.dart';
 
 @lazySingleton
 class LightService {
-  final _prefsService = locator<SharedPreferences>();
-  final _deviceService = locator<DeviceService>();
+  final SharedPreferences? _prefsService = locator<SharedPreferences>();
+  final DeviceService? _deviceService = locator<DeviceService>();
 
   final Client _client = Client();
 
@@ -29,17 +29,15 @@ class LightService {
   }
 
   Future<bool> linkBridge(BridgeModel bridgeModel) async {
-    final bridge = Bridge(_client, bridgeModel.ip);
+    final bridge = Bridge(_client, bridgeModel.ip!);
 
     final auth = await _createUser(bridge);
-    if (auth != null) {
-      bridgeModel
-        ..auth = auth
-        ..state = Connection.connected;
-      final bridges = <BridgeModel>[bridgeModel];
-      final json = BridgeModel.encode(bridges);
-      await _prefsService.setString(kPrefKeyHueBridges, json);
-    }
+    bridgeModel
+      ..auth = auth
+      ..state = Connection.connected;
+    final bridges = <BridgeModel>[bridgeModel];
+    final json = BridgeModel.encode(bridges);
+    await _prefsService!.setString(kPrefKeyHueBridges, json);
 
     return false;
   }
@@ -49,19 +47,19 @@ class LightService {
       return Connection.unsaved;
     }
 
-    final bridge = Bridge(_client, bridgeModel.ip, bridgeModel.auth);
-    var config = await bridge.configuration();
+    final bridge = Bridge(_client, bridgeModel.ip!, bridgeModel.auth!);
+    final config = await bridge.configuration();
 
-    if (config == null || config.whitelist == null) {
+    if (config.whitelist == null) {
       return Connection.failed;
     }
 
     return Connection.connected;
   }
 
-  Future<String> _createUser(Bridge bridge) async {
+  Future<String?> _createUser(Bridge bridge) async {
     final username =
-        '$kHueBridgeUsername#${_deviceService.deviceManufacturer} ${_deviceService.deviceModel}';
+        '$kHueBridgeUsername#${_deviceService!.deviceManufacturer} ${_deviceService!.deviceModel}';
     try {
       final whiteListItem = await bridge.createUser(username);
       return whiteListItem.username;
@@ -71,7 +69,7 @@ class LightService {
   }
 
   Future<List<LightGroup>> getRooms(BridgeModel bridgeModel) async {
-    final bridge = Bridge(_client, bridgeModel.ip, bridgeModel.auth);
+    final bridge = Bridge(_client, bridgeModel.ip!, bridgeModel.auth!);
     final groups = await bridge.groups();
     return groups
         .where((e) => e.type?.toLowerCase() != 'entertainment')
@@ -79,12 +77,12 @@ class LightService {
             id: e.id.toString(),
             className: e.className,
             name: e.name,
-            numberOfLights: e.lightIds.length))
+            numberOfLights: e.lightIds?.length))
         .toList();
   }
 
   Future<void> toggleLights(final bool enabled) async {
-    final savedBridgesJson = _prefsService.getString(kPrefKeyHueBridges);
+    final savedBridgesJson = _prefsService!.getString(kPrefKeyHueBridges);
 
     if (savedBridgesJson != null) {
       final savedBridges = BridgeModel.decode(savedBridgesJson);
@@ -93,16 +91,17 @@ class LightService {
         final connectionState = await getConnectionState(bridgeModel);
 
         if (connectionState == Connection.connected) {
-          final bridge = Bridge(_client, bridgeModel.ip, bridgeModel.auth);
+          final bridge = Bridge(_client, bridgeModel.ip!, bridgeModel.auth!);
           final groups = await bridge.groups();
 
           bridgeModel.groups.forEach((group) {
-            if (group.actionEnabled) {
+            if (group.actionEnabled!) {
               final foundGroup = groups
                   .singleWhere((element) => element.id.toString() == group.id);
-              if (foundGroup != null && foundGroup.state.anyOn) {
+
+              if (foundGroup.state?.anyOn ?? true) {
                 foundGroup.groupLights
-                    .where((light) => light.state.on)
+                    ?.where((light) => light.state?.on ?? true)
                     .forEach((shiningLight) {
                   bridge.updateLightState(shiningLight.rebuild((e) => e
                     ..state = lightStateForColorOnly(shiningLight)
@@ -116,4 +115,19 @@ class LightService {
       });
     }
   }
+}
+
+LightState lightStateForColorOnly(Light _light) {
+  LightState state;
+  if (_light.state?.colorMode == 'xy') {
+    state = LightState((b) => b..xy = _light.state?.xy?.toBuilder());
+  } else if (_light.state?.colorMode == 'ct') {
+    state = LightState((b) => b..ct = _light.state?.ct);
+  } else {
+    state = LightState((b) => b
+      ..hue = _light.state?.hue
+      ..saturation = _light.state?.saturation
+      ..brightness = _light.state?.brightness);
+  }
+  return state;
 }
