@@ -1,7 +1,8 @@
+import 'package:collection/collection.dart' show IterableExtension;
 import 'package:hue_dart/hue_dart.dart';
 import 'package:logger/logger.dart';
 import 'package:shared_preferences/shared_preferences.dart';
-import 'package:sleep_timer/app/auto_router.gr.dart';
+import 'package:sleep_timer/app/app.router.dart';
 import 'package:sleep_timer/app/locator.dart';
 import 'package:sleep_timer/app/logger.util.dart';
 import 'package:sleep_timer/common/constants.dart';
@@ -12,31 +13,30 @@ import 'package:stacked_services/stacked_services.dart';
 
 class BridgeLinkViewModel extends FutureViewModel {
   final Logger log = getLogger();
-  final _navigationService = locator<NavigationService>();
-  final _prefsService = locator<SharedPreferences>();
-  final _lightService = locator<LightService>();
+  final NavigationService? _navigationService = locator<NavigationService>();
+  final SharedPreferences? _prefsService = locator<SharedPreferences>();
+  final LightService? _lightService = locator<LightService>();
 
   String connectionError = '';
 
   @override
   Future futureToRun() async {
     try {
-      final bridges = await _lightService.findBridges();
+      final bridges = await _lightService!.findBridges();
 
-      final savedBridgesJson = _prefsService.getString(kPrefKeyHueBridges);
+      final savedBridgesJson = _prefsService!.getString(kPrefKeyHueBridges);
 
       if (savedBridgesJson == null) return bridges;
 
       final savedBridges = BridgeModel.decode(savedBridgesJson);
       await Future.forEach(bridges, (BridgeModel element) async {
-        final savedEntry = savedBridges.firstWhere(
-            (element) => element.id == element.id,
-            orElse: () => null);
+        final savedEntry = savedBridges.firstWhereOrNull(
+            (element) => element.id == element.id);
 
         if (savedEntry != null) {
           element
             ..auth = savedEntry.auth
-            ..state = await _lightService.getConnectionState(element);
+            ..state = await _lightService!.getConnectionState(element);
         }
       });
 
@@ -53,12 +53,12 @@ class BridgeLinkViewModel extends FutureViewModel {
 
   Future<void> linkBridge(final BridgeModel bridgeModel) async {
     try {
-      await _lightService.linkBridge(bridgeModel);
+      await _lightService!.linkBridge(bridgeModel);
       _resetError();
 
       navigateBackToLights();
     } on BridgeException catch (error) {
-      connectionError = error.description.toUpperCase();
+      connectionError = error.description?.toUpperCase() ?? 'ConnectionError';
       notifyListeners();
       log.e(error);
     }
@@ -67,17 +67,16 @@ class BridgeLinkViewModel extends FutureViewModel {
   }
 
   Future<bool> connect(final BridgeModel bridgeModel) async {
-    final success = await _lightService.linkBridge(bridgeModel);
+    final success = await _lightService!.linkBridge(bridgeModel);
 
     if (success) {
-      final savedBridgesJson = _prefsService.getString(kPrefKeyHueBridges);
+      final savedBridgesJson = _prefsService!.getString(kPrefKeyHueBridges);
       var savedBridges = <BridgeModel>[];
 
       if (savedBridgesJson != null) {
         savedBridges = BridgeModel.decode(savedBridgesJson);
-        final savedEntry = savedBridges.firstWhere(
-            (element) => element.id == bridgeModel.id,
-            orElse: () => null);
+        final savedEntry = savedBridges.firstWhereOrNull(
+            (element) => element.id == bridgeModel.id);
 
         // match with saved bridge --> update
         if (savedEntry != null) {
@@ -92,7 +91,7 @@ class BridgeLinkViewModel extends FutureViewModel {
       }
 
       // save updates
-      await _prefsService.setString(
+      await _prefsService!.setString(
           kPrefKeyHueBridges, BridgeModel.encode(savedBridges));
     }
 
@@ -100,18 +99,18 @@ class BridgeLinkViewModel extends FutureViewModel {
   }
 
   Future<void> remove(final BridgeModel bridgeModel) async {
-    await _prefsService.remove(kPrefKeyHueBridges);
-    _navigationService.back();
+    await _prefsService!.remove(kPrefKeyHueBridges);
+    _navigationService!.back();
     await initialise();
   }
 
   void navigateBackToLights() {
-    _navigationService
+    _navigationService!
         .popUntil((route) => route.settings.name == Routes.lightGroupView);
   }
 
   void cancelDialog() {
-    _navigationService.back();
+    _navigationService!.back();
     _resetError();
   }
 }
